+++ client/src/common/styles/buttonStyles.ts
@@ -1,0 +1,25 @@
+export const submitButtonStyles = {
+    background: 'linear-gradient(45deg, #1976d2, #9c27b0)',
+    color: 'white',
+    fontWeight: 'bold',
+    borderRadius: '20px',
+    padding: '10px 60px',
+    textTransform: 'none',
+    mx: 1,
+    '&:hover': {
+        background: 'linear-gradient(45deg, #1565c0, #7b1fa2)',
+    },
+}
+
+export const cancelButtonStyles = {
+    color: '#1976d2',
+    fontWeight: 'bold',
+    borderRadius: '20px',
+    border: '2px solid #1976d2',
+    padding: '10px 60px',
+    textTransform: 'none',
+    '&:hover': {
+        borderColor: '#1565c0',
+        color: '#1565c0',
+    },
+}
