+++ client/src/common/styles/loginTextStyles.ts
@@ -1,0 +1,12 @@
+import { Theme } from '@mui/material'
+
+export const loginTextStyles = {
+    ml: 2,
+    color: (theme: Theme) => theme.palette.common.white,
+    fontWeight: 'bold',
+    fontSize: '1rem',
+    letterSpacing: 1.0,
+    textAlign: 'center',
+    padding: '5px 15px',
+    borderRadius: '20px',
+}
